use staticLibary

+++ Puppet/src/Puppet/Platform/Windows/WindowsWindow.cpp
@@ -121,13 +121,13 @@ namespace Puppet {
 		glfwSetScrollCallback(m_Window, [](GLFWwindow* window, double xoffset, double yoffset)
 		{
 			WindowData* data = static_cast<WindowData*>(glfwGetWindowUserPointer(window));
-			MouseScrolledEvent event(xoffset, yoffset);
+			MouseScrolledEvent event((float)xoffset, (float)yoffset);
 			data->EventCallback(event);
 		});
 		glfwSetCursorPosCallback(m_Window, [](GLFWwindow* window, double xpos, double ypos)
 		{
 			WindowData* data = static_cast<WindowData*>(glfwGetWindowUserPointer(window));
-			MouseMovedEvent event(xpos, ypos);
+			MouseMovedEvent event((float)xpos, (float)ypos);
 			data->EventCallback(event);
 		});
 		glfwSetCharCallback(m_Window, [](GLFWwindow* window, unsigned int keycode)
